+++ resources/js/Components/VideoPlayer/Chat/FullPageChat.vue
@@ -1,9 +1,5 @@
 <template>
     <div>
-        <div class="flex flex-col ">
-            <div class="text-3xl font-semibold">Conversation</div>
-        </div>
-        <div class="italic mb-10">The newest message is at the bottom.</div>
         <div class="absolute">
             <div>
                 <chat-messages />
@@ -12,7 +8,7 @@
                 <input-message
                     :user="props.user"
                     class="chatFullPageInput"
-                    :class="{ 'text-gray-100': !chatStore.inputTooLong, 'text-red-600': chatStore.inputTooLong }"
+                    :class="[{ 'text-gray-100': !chatStore.inputTooLong, 'text-red-600': chatStore.inputTooLong }, {'isMobileComputed': userStore.isMobile}]"
                 />
             </div>
         </div>
@@ -24,9 +20,11 @@ import InputMessage from "@/Components/VideoPlayer/Chat/FullPageChatInput"
 import ChatMessages from "@/Components/VideoPlayer/Chat/FullPageChatMessages"
 import { useVideoPlayerStore } from "@/Stores/VideoPlayerStore";
 import { useChatStore } from "@/Stores/ChatStore";
+import { useUserStore } from "@/Stores/UserStore";
 
 let videoPlayer = useVideoPlayerStore()
 let chatStore = useChatStore()
+let userStore = useUserStore()
 
 let props = defineProps({
     user: Object,